+++ src/baryo_fhck/TransportEquations.cpp
@@ -523,7 +523,7 @@ void TransportEquations::Equations(const double &z,
   VecDoub BosonMasses(nBosons);
 
   // Fermions
-  for (int fermion = 0; fermion < nFermions; fermion++)
+  for (size_t fermion = 0; fermion < nFermions; fermion++)
   {
 
     double m2, m2prime, thetaprime, theta2prime;
@@ -552,7 +552,7 @@ void TransportEquations::Equations(const double &z,
   }
 
   // Bosons
-  for (int boson = 0; boson < nBosons; boson++)
+  for (size_t boson = 0; boson < nBosons; boson++)
   {
     double m2, m2prime;
     // Calculate the boson mass
@@ -579,14 +579,14 @@ void TransportEquations::Equations(const double &z,
       CalculateCollisionMatrix(mW, FermionMasses, BosonMasses);
 
   // Calculate M = A^-1 * Gamma ( = deltaC - m2'B)
-  for (int i = 0; i < 2 * (nBosons + nFermions); i++)
-    for (int j = 0; j < 2 * (nBosons + nFermions); j++)
-      for (int l = 0; l < 2 * (nBosons + nFermions); l++)
+  for (size_t i = 0; i < 2 * (nBosons + nFermions); i++)
+    for (size_t j = 0; j < 2 * (nBosons + nFermions); j++)
+      for (size_t l = 0; l < 2 * (nBosons + nFermions); l++)
         Mtilde[i][j] += Ainverse[i][l] * (CollisiontMatrix[l][j] - m2B[l][j]);
 
   // Calculate Stilde = A^-1 * S
-  for (int i = 0; i < 2 * (nBosons + nFermions); i++)
-    for (int j = 0; j < 2 * (nFermions); j++)
+  for (size_t i = 0; i < 2 * (nBosons + nFermions); i++)
+    for (size_t j = 0; j < 2 * (nFermions); j++)
       Stilde[i] += Ainverse[i][j] * S[j];
 }
